Feature/ main world/main world tile (#37)
* zonebase 를 zonebasic 의 부모클래스

* 로그 추가 및 Grid[Add/Release] 시 매개변수로 key 값 주게 변

* 일단 MainWorld 을 Tile 로 분리후 Grid 에 등록 까지

* commit 안된 항목 커

* 로그인 시 Player Ref 관련 이슈 수

* Teleport 조건 수

* Player Login Ref 관련 오류 수정
로그인시 Handoff 과정에서 Leave 가 먼저 실행 되서 문제 발생 이슈

* 왜 스테이징 된 항목된 커밋이 되는 옵션만 있는걸까 비쥬얼 스튜디오 에는?

* 디버그 를 위해 추가한 변수 삭제

+++ GameServer/CUtill/CBinFileManager.cpp
@@ -20,7 +20,7 @@ bool CBinFileManager::ReadBinZoneIdxFile(const char* filepath)
 		return false;
 	
 	const std::vector<IDX> vecZoneIdx = bin.GetZoneIdxVector();
-	int Loop = vecZoneIdx.size();
+	int Loop = static_cast<int>(vecZoneIdx.size());
 	for (int i = 0; i < Loop; i++)
 	{
 		const IDX& data = vecZoneIdx[i];
@@ -61,7 +61,7 @@ bool CBinFileManager::ReadBinZoneFile(const char* filepath)
 	
 	std::vector<BinPortal> vecPortal = bin.GetPortalVector();
 	{
-		int Loop = vecPortal.size();
+		int Loop = static_cast<int>(vecPortal.size());
 		for (int i = 0; i < Loop; i++)
 		{
 			st_Portal portal;
@@ -82,7 +82,7 @@ bool CBinFileManager::ReadBinZoneFile(const char* filepath)
 
 	std::vector<BinSpawnPoint> vecSpawnPoint = bin.GetSpawnPointVector();
 	{
-		int Loop = vecSpawnPoint.size();
+		int Loop = static_cast<int>(vecSpawnPoint.size());
 		for (int i = 0; i < Loop; i++)
 		{
 			st_SpawnPoint spawn;
@@ -102,7 +102,7 @@ bool CBinFileManager::ReadBinZoneFile(const char* filepath)
 
 	std::vector<BinTriggerVolume> vecTriggerVolume = bin.GetTriggerVolumeVector();
 	{
-		int Loop = vecTriggerVolume.size();
+		int Loop = static_cast<int>(vecTriggerVolume.size());
 		for (int i = 0; i < Loop; i++)
 		{
 			st_TriggerVolume trigger;
@@ -126,7 +126,7 @@ bool CBinFileManager::ReadBinZoneFile(const char* filepath)
 			int triggerId = iter->first;
 			std::vector<BinTriggerVolumeParam> vecParam = iter->second;
 			std::vector<st_TriggerVolumeParam> vecStParam;
-			int Loop = vecParam.size();
+			int Loop = static_cast<int>(vecParam.size());
 			for (int i = 0; i < Loop; i++)
 			{
 				st_TriggerVolumeParam stParam;
